Created file structure for the resolvers and broke them up into separate folders and files.
